block_tree_pare added
block_tree_pare cleans up a block tree, freeing all unused blocks and
releasing them back to the file system. It is intended to be used after
a file was opened for writing, written to, and then closed. Since the
newly written file may not be as large as the previous one, it will most
likely have blocks in its tree which are not being used. These are the
blocks intended to be freed.